+++ CHANGELOG.md
@@ -15,13 +15,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Use instance based RustSBI interface, with separate functions for legacy stdio
 - Update sbi-testing to version 0.0.1
+- Use crate fast-trap version 0.0.1 in rustsbi-qemu for trap handling
 - Use crate rcore-console version 0.0.0 in rustsbi-qemu and test-kernel for `print!` and `println!`
 - Use crate aclint version 0.0.0 in rustsbi-qemu for aclint structs
 - Use crate os-xtask-utils version 0.0.0 in xtask builder
+- Use crate sifive-test-device version 0.0.0 instead of qemu-exit
+- Use `wfi` for suspend and stop without enable mie
 
 ### Fixed
 
 - Xtask will now print error when system does not have qemu installed
+- Fix dtb parsing for qemu 7.2
 
 ## [0.1.1] - 2022-03-23